TrashNet-Classification/requirements.txt
tqdm==4.66.5
imutils==0.5.4
numpy==1.26.4
pandas==2.0.3
pillow==10.4.0
matplotlib==3.7.3
seaborn==0.11.0
albumentations==1.4.1
opencv-python==4.10.0.84
tensorflow==2.15.1
keras==2.15.1
scikit-learn==1.2.2
wandb==0.19.1